pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = ""
authors = ["Caio Wingeter <castilhocaio1164@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
anthropic = "^0.29.0"
gitpython = "^3.1.43"
loguru = "^0.7.2"
prompt-toolkit = "^3.0.47"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = "/home/caio/.cache/pypoetry/virtualenvs/ai-2g66xYS9-py3.12"
